Don't load files from a custom folder
The previous version enabled the iTerm2 setting that lets you load a
preference file from a custom location. This custom location was a
specific location on my computer, which obviously errored out when we
used this on another machine. This setting repo is purely for
Puppet-Boxen setups, so we let it handle getting the preference files
into the right place via sym links. We don't need this setting in iterm.